Migrate legacy api endpoints to tRPC II (#3865) (#3991)
* Migrate legacy api endpoints to tRPC II (#3865)

* fix: delete api route /api/availability/day

* fix: delete api route /api/availability/eventtype (#3573)

* fix: delete api route /api/event-type (#3573)

* migrate /api/book/confirm to viewer.bookings.confirm

* Merge branch 'main' into pr/3865


* Import fixes

* Import fixes
