+++ apps/web/calendso.yaml
@@ -277,179 +277,6 @@ paths:
                   type: string
                 integration:
                   type: string
-  /api/availability/day:
-    patch:
-      description: Updates the start and end times for a user's availability.
-      summary: Updates the user's start and end times
-      tags:
-        - Availability
-      responses:
-        "200":
-          description: OK
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  message:
-                    type: string
-      requestBody:
-        content:
-          application/json:
-            schema:
-              type: object
-              properties:
-                startMins:
-                  type: string
-                endMins:
-                  type: string
-                bufferMins:
-                  type: string
-        description: ""
-  /api/availability/eventtype:
-    post:
-      description: Adds a new event type for the user.
-      summary: Adds a new event type
-      tags:
-        - Availability
-      requestBody:
-        content:
-          application/json:
-            schema:
-              type: object
-              properties:
-                title:
-                  type: string
-                slug:
-                  type: string
-                description:
-                  type: string
-                length:
-                  type: string
-                hidden:
-                  type: boolean
-                requiresConfirmation:
-                  type: boolean
-                locations:
-                  type: array
-                  items: {}
-                eventName:
-                  type: string
-                customInputs:
-                  type: array
-                  items: {}
-      responses:
-        "200":
-          description: OK
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  eventType:
-                    type: object
-                    properties:
-                      title:
-                        type: string
-                      slug:
-                        type: string
-                      description:
-                        type: string
-                      length:
-                        type: string
-                      hidden:
-                        type: boolean
-                      requiresConfirmation:
-                        type: boolean
-                      locations:
-                        type: array
-                        items: {}
-                      eventName:
-                        type: string
-                      customInputs:
-                        type: array
-                        items: {}
-        "500":
-          description: Internal Server Error
-    patch:
-      description: Updates an event type for the user.
-      summary: Updates an event type
-      tags:
-        - Availability
-      requestBody:
-        content:
-          application/json:
-            schema:
-              type: object
-              properties:
-                title:
-                  type: string
-                slug:
-                  type: string
-                description:
-                  type: string
-                length:
-                  type: string
-                hidden:
-                  type: boolean
-                requiresConfirmation:
-                  type: boolean
-                locations:
-                  type: array
-                  items: {}
-                eventName:
-                  type: string
-                customInputs:
-                  type: array
-                  items: {}
-      responses:
-        "200":
-          description: OK
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  eventType:
-                    type: object
-                    properties:
-                      title:
-                        type: string
-                      slug:
-                        type: string
-                      description:
-                        type: string
-                      length:
-                        type: string
-                      hidden:
-                        type: boolean
-                      requiresConfirmation:
-                        type: boolean
-                      locations:
-                        type: array
-                        items: {}
-                      eventName:
-                        type: string
-                      customInputs:
-                        type: array
-                        items: {}
-        "500":
-          description: Internal Server Error
-    delete:
-      description: Deletes an event type for the user.
-      summary: Deletes an event type
-      tags:
-        - Availability
-      responses:
-        "200":
-          description: OK
-          content:
-            application/json:
-              schema:
-                type: object
-                properties: {}
-        "500":
-          description: Internal Server Error
   "/api/book/event":
     post:
       description: Creates a booking in the user's calendar.
