+++ apps/web/lib/webhooks/subscriptions.tsx
@@ -8,6 +8,7 @@ export type GetSubscriberOptions = {
   triggerEvent: WebhookTriggerEvents;
 };
 
+/** @deprecated use `packages/lib/webhooks/subscriptions.tsx` */
 const getWebhooks = async (options: GetSubscriberOptions) => {
   const { userId, eventTypeId } = options;
   const allWebhooks = await prisma.webhook.findMany({
